Moved Header, add CSS and made header 100% and added icons and color
