+++ client/src/components/Header/Header.js
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
+import './Header.css';
 
 class Header extends Component {
     renderContent() {
@@ -35,8 +36,8 @@ class Header extends Component {
                 <div className="nav-wrapper">
                     <Link 
                         to={user ? '/dashboard' : '/'}
-                        className="left brand-logo">
-                        Music Media
+                        className="left">
+                        <i className="material-icons library-music">library_music</i>
                     </Link>
                     <ul className="right">
                         {this.renderContent()}
